changed u128 to LookupId

+++ crates/core/executor/src/syscalls/precompiles/fptower/fp2_addsub.rs
@@ -63,7 +63,7 @@ impl<P: FpOpField> Syscall for Fp2AddSubSyscall<P> {
         result.resize(num_words, 0);
         let x_memory_records = rt.mw_slice(x_ptr, &result);
 
-        let lookup_id = rt.syscall_lookup_id as usize;
+        let lookup_id = rt.syscall_lookup_id;
         let shard = rt.current_shard();
         let channel = rt.current_channel();
         let op = self.op;
